# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/properties": {
            "get": {
                "tags": [
                    "appartement"
                ],
                "summary": "Voir les appartements",
                "description": "Voir tous les appartements",
                "operationId": "getInventory",
                "responses": {
                    "200": {
                        "description": "réussite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/property/add": {
            "put": {
                "tags": [
                    "appartement"
                ],
                "summary": "Ajouter un nouvel appartement.",
                "operationId": "updatePet",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Appartement"
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Appartement non trouver"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:appartements",
                            "read:appartements"
                        ]
                    }
                ]
            }
        },
        "/property/modify/{Id}": {
            "post": {
                "tags": [
                    "appartement"
                ],
                "summary": "Update an existing appartement.",
                "operationId": "update",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Appartement"
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Appartement not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:appartements",
                            "read:appartements"
                        ]
                    }
                ]
            }
        },
        "/property/delete/{Id}": {
            "delete": {
                "tags": [
                    "appartement"
                ],
                "summary": "Deletes a appartement",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Appartement id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Appartement not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:appartements",
                            "read:appartements"
                        ]
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}